+++ apps/web/app/components/ProductAccordion/ProductAccordion.vue
@@ -1,18 +1,6 @@
 <template>
   <div data-testid="product-accordion">
-    <UiAccordionItem
-      v-if="productGetters.getDescription(product)?.length"
-      v-model="productDetailsOpen"
-      summary-class="md:rounded-md w-full hover:bg-neutral-100 py-2 pl-4 pr-3 flex justify-between items-center select-none"
-    >
-      <template #summary>
-        <h2 class="font-bold text-lg leading-6 md:text-2xl">
-          {{ t('productDetails') }}
-        </h2>
-      </template>
-      <div class="no-preflight" v-html="productGetters.getDescription(product)" />
-    </UiAccordionItem>
-    <UiDivider v-if="productDetailsOpen && productGetters.getDescription(product)?.length" class="mb-2 mt-2" />
+    <BlocksProductDetails :product="product" />
     <UiAccordionItem
       v-if="productGetters.getTechnicalData(product)?.length"
       v-model="technicalDataOpen"
@@ -30,6 +18,7 @@
 </template>
 
 <script setup lang="ts">
+import { BlocksProductDetails } from '#components';
 import { productGetters } from '@plentymarkets/shop-api';
 import type { ProductAccordionPropsType } from '~/components/ProductAccordion/types';
 
@@ -38,6 +27,5 @@ const props = defineProps<ProductAccordionPropsType>();
 const { product } = toRefs(props);
 const { t } = useI18n();
 
-const productDetailsOpen = ref(true);
 const technicalDataOpen = ref(false);
 </script>
