+++ apps/web/app/components/blocks/ProductDetails/types.ts
@@ -1,0 +1,4 @@
+import type { Product } from '@plentymarkets/shop-api';
+export type ProductDetailsProps = {
+    product: Product
+};
